Move input file handling to FilePrep. Change XYZ values to be quantized to output scaling.
