+++ epf/Grid.hpp
@@ -15,6 +15,7 @@
 
 #include <pdal/util/Bounds.hpp>
 
+#include "../untwine/Common.hpp"
 #include "../untwine/VoxelKey.hpp"
 
 namespace untwine
@@ -25,32 +26,24 @@ namespace epf
 class Grid
 {
 public:
-    Grid() : m_gridSize(-1), m_maxLevel(-1), m_millionPoints(0), m_cubic(true)
-    {}
+    Grid(const pdal::BOX3D& bounds, uint64_t numPoints, int level, bool cubic) :
+        m_bounds(bounds)
+    {
+        if (level == -1)
+            level = calcLevel(numPoints, cubic);
+        resetLevel(level);
+    }
 
-    void expand(const pdal::BOX3D& bounds, size_t points);
-    int calcLevel();
-    void resetLevel(int level);
     VoxelKey key(double x, double y, double z) const;
-    pdal::BOX3D processingBounds() const
-        { return m_cubic ? m_cubicBounds : m_bounds; }
-    pdal::BOX3D cubicBounds() const
-        { return m_cubicBounds; }
-    pdal::BOX3D conformingBounds() const
-        { return m_bounds; }
-
+    int calcLevel(uint64_t numPoints, bool cubic) const;
+    void resetLevel(int level);
     int maxLevel() const
         { return m_maxLevel; }
-    void setCubic(bool cubic)
-        { m_cubic = cubic; }
 
 private:
+    pdal::BOX3D m_bounds;
     int m_gridSize;
     int m_maxLevel;
-    pdal::BOX3D m_bounds;
-    pdal::BOX3D m_cubicBounds;
-    size_t m_millionPoints;
-    bool m_cubic;
     double m_xsize;
     double m_ysize;
     double m_zsize;
